Refactor and re-package dto classes
