+++ src/main/java/com/bytebreeze/quickdrop/controller/UserController.java
@@ -1,6 +1,6 @@
 package com.bytebreeze.quickdrop.controller;
 
-import com.bytebreeze.quickdrop.dto.UserProfileUpdateDto;
+import com.bytebreeze.quickdrop.dto.request.UserProfileUpdateDto;
 import com.bytebreeze.quickdrop.dto.request.ParcelBookingRequestDTO;
 import com.bytebreeze.quickdrop.model.Parcel;
 import com.bytebreeze.quickdrop.repository.ProductCategoryRepository;
