+++ src/main/java/com/bytebreeze/quickdrop/dto/request/RiderRegistrationRequestDTO.java
@@ -1,4 +1,4 @@
-package com.bytebreeze.quickdrop.dto;
+package com.bytebreeze.quickdrop.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotEmpty;
