+++ src/main/java/com/bytebreeze/quickdrop/dto/request/UserProfileUpdateDto.java
@@ -1,4 +1,4 @@
-package com.bytebreeze.quickdrop.dto;
+package com.bytebreeze.quickdrop.dto.request;
 
 import jakarta.validation.constraints.NotEmpty;
 import lombok.Data;
