+++ src/main/java/com/bytebreeze/quickdrop/mapper/OnboardRiderMapper.java
@@ -1,6 +1,6 @@
 package com.bytebreeze.quickdrop.mapper;
 
-import com.bytebreeze.quickdrop.dto.RiderOnboardingDTO;
+import com.bytebreeze.quickdrop.dto.request.RiderOnboardingDTO;
 import com.bytebreeze.quickdrop.model.Rider;
 import org.mapstruct.Mapper;
 import org.springframework.stereotype.Component;
