+++ src/main/java/com/bytebreeze/quickdrop/mapper/RegisterRiderMapper.java
@@ -1,6 +1,6 @@
 package com.bytebreeze.quickdrop.mapper;
 
-import com.bytebreeze.quickdrop.dto.RiderRegistrationRequestDTO;
+import com.bytebreeze.quickdrop.dto.request.RiderRegistrationRequestDTO;
 import com.bytebreeze.quickdrop.model.Rider;
 import org.mapstruct.Mapper;
 import org.springframework.stereotype.Component;
