+++ src/main/java/com/bytebreeze/quickdrop/service/UserService.java
@@ -1,7 +1,7 @@
 package com.bytebreeze.quickdrop.service;
 
-import com.bytebreeze.quickdrop.dto.UserProfileUpdateDto;
-import com.bytebreeze.quickdrop.dto.UserRegistrationRequestDTO;
+import com.bytebreeze.quickdrop.dto.request.UserProfileUpdateDto;
+import com.bytebreeze.quickdrop.dto.request.UserRegistrationRequestDTO;
 import com.bytebreeze.quickdrop.enums.Role;
 import com.bytebreeze.quickdrop.exception.custom.AlreadyExistsException;
 import com.bytebreeze.quickdrop.exception.custom.UserNotFoundException;
