+++ src/test/java/com/bytebreeze/quickdrop/controller/AuthControllerTest.java
@@ -5,7 +5,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.*;
 
-import com.bytebreeze.quickdrop.dto.UserRegistrationRequestDTO;
+import com.bytebreeze.quickdrop.dto.request.UserRegistrationRequestDTO;
 import com.bytebreeze.quickdrop.service.UserService;
 import com.bytebreeze.quickdrop.util.AuthUtil;
 import java.util.List;
