+++ src/test/java/com/bytebreeze/quickdrop/util/ProfileSettingUtilTest.java
@@ -3,7 +3,7 @@ package com.bytebreeze.quickdrop.util;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
-import com.bytebreeze.quickdrop.dto.UserProfileUpdateDto;
+import com.bytebreeze.quickdrop.dto.request.UserProfileUpdateDto;
 import com.bytebreeze.quickdrop.service.UserService;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
